+++ source/audiolib/src/xmp.cpp
@@ -18,13 +18,13 @@ typedef struct {
     int time;
 } xmp_data;
 
-int32_t MV_GetXMPPosition(VoiceNode *voice)
+int MV_GetXMPPosition(VoiceNode *voice)
 {
     auto xmpd = (xmp_data *)voice->rawdataptr;
     return xmpd->time;
 }
 
-void MV_SetXMPPosition(VoiceNode *voice, int32_t position)
+void MV_SetXMPPosition(VoiceNode *voice, int position)
 {
     auto xmpd = (xmp_data *)voice->rawdataptr;
     xmp_seek_time(xmpd->context, position);
@@ -64,13 +64,13 @@ static playbackstatus MV_GetNextXMPBlock(VoiceNode *voice)
     return KeepPlaying;
 }
 
-int32_t MV_PlayXMP3D(char *ptr, uint32_t length, int32_t loophow, int32_t pitchoffset, int32_t angle, int32_t distance, int32_t priority, float volume, intptr_t callbackval)
+int MV_PlayXMP3D(char *ptr, uint32_t length, int loophow, int pitchoffset, int angle, int distance, int priority, float volume, uint32_t callbackval)
 {
-    int32_t left;
-    int32_t right;
-    int32_t mid;
-    int32_t vol;
-    int32_t status;
+    int left;
+    int right;
+    int mid;
+    int vol;
+    int status;
 
     if (!MV_Installed)
         return MV_SetErrorCode(MV_NotInstalled);
@@ -95,7 +95,7 @@ int32_t MV_PlayXMP3D(char *ptr, uint32_t length, int32_t loophow, int32_t pitcho
     return status;
 }
 
-int32_t MV_PlayXMP(char *ptr, uint32_t length, int32_t loopstart, int32_t loopend, int32_t pitchoffset, int32_t vol, int32_t left, int32_t right, int32_t priority, float volume, intptr_t callbackval)
+int MV_PlayXMP(char *ptr, uint32_t length, int loopstart, int loopend, int pitchoffset, int vol, int left, int right, int priority, float volume, uint32_t callbackval)
 {
     VoiceNode   *voice;
     xmp_data * xmpd = 0;
@@ -199,8 +199,8 @@ void MV_ReleaseXMPVoice(VoiceNode * voice)
 
 static char const NoXMP[] = "MV_PlayXMP: libxmp-lite support not included in this binary.\n";
 
-int32_t MV_PlayXMP(char *ptr, uint32_t ptrlength, int32_t loopstart, int32_t loopend, int32_t pitchoffset, int32_t vol,
-                   int32_t left, int32_t right, int32_t priority, float volume, intptr_t callbackval)
+int MV_PlayXMP(char *ptr, uint32_t ptrlength, int loopstart, int loopend, int pitchoffset, int vol,
+                   int left, int right, int priority, float volume, uint32_t callbackval)
 {
     UNREFERENCED_PARAMETER(ptr);
     UNREFERENCED_PARAMETER(ptrlength);
@@ -218,8 +218,8 @@ int32_t MV_PlayXMP(char *ptr, uint32_t ptrlength, int32_t loopstart, int32_t loo
     return -1;
 }
 
-int32_t MV_PlayXMP3D(char *ptr, uint32_t ptrlength, int32_t loophow, int32_t pitchoffset, int32_t angle,
-                     int32_t distance, int32_t priority, float volume, intptr_t callbackval)
+int MV_PlayXMP3D(char *ptr, uint32_t ptrlength, int loophow, int pitchoffset, int angle,
+                     int distance, int priority, float volume, uint32_t callbackval)
 {
     UNREFERENCED_PARAMETER(ptr);
     UNREFERENCED_PARAMETER(ptrlength);
